feat: deprecated_tool lint rule, integer `number` samples, corrected-α in fix verdicts
From the real-server scenarios (docs/examples/mcp-server-git, server-filesystem):
- server-filesystem's read_file says "DEPRECATED: Use read_text_file instead"
  and Sonnet 5 obeys — 0/10 by design. `scan` now flags any tool whose
  description calls itself deprecated; it's the first static rule that fires
  on a mature public server.
- zod-generated schemas type counts as `number`; the sampler produced
  "read the first 76.22 lines" and the solvability check flagged it. Sample
  integers — a valid number that reads like a real request.
- "not significant after correction" now prints p vs the corrected α, so a
  12-proposal run at n=10 (α=0.004, unreachable) explains itself.

+++ tests/test_schema_sampler.py
@@ -128,7 +128,7 @@ def test_sample_arguments_generates_integer_number_and_boolean():
     }
     result = sample_arguments(schema, seed=1)
     assert isinstance(result["count"], int)
-    assert isinstance(result["score"], float)
+    assert isinstance(result["score"], int)  # integers are valid numbers and read naturally in tasks
     assert isinstance(result["active"], bool)
